added struct for a structured map

+++ src/constants.rs
@@ -1,2 +1,2 @@
-pub const FIELDWIDTH : u32 = 600;
-pub const FIELDHEIGHT: u32 = 400;
+pub const FIELDWIDTH : usize = 600;
+pub const FIELDHEIGHT: usize = 400;
